Created a survey repository and survey controllers
Created survey repository and Controllers. There some bugs in the controller
