+++ src/main/java/minisurveymonkey/accessingdatajpa/SurveyRepository.java
@@ -1,0 +1,10 @@
+package minisurveymonkey.accessingdatajpa;
+
+import minisurveymonkey.accessingdatajpa.Survey;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+
+@RepositoryRestResource(collectionResourceRel = "survey", path = "survey")
+public interface  SurveyRepository extends CrudRepository<Survey, Long> {
+    Survey findById(long id);
+}
