+++ src/main/resources/templates/creationView.html
@@ -1,0 +1,47 @@
+<!DOCTYPE HTML>
+<html>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+<head>
+    <title>Mini Survey Monkey</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.0/dist/css/bootstrap.min.css"
+          integrity="sha384-B0vP5xmATw1+K9KRQjQERJvTumQW0nPEzvF6L/Z6nronJ3oUOFUFpCjEUQouq2+l" crossorigin="anonymous">
+</head>
+<body>
+<!-- class="container mt-5 pt-5 " -->
+<div class="container mt-5 pt-5 ">
+    <div class="row ">
+        <div class="col col-6">
+            <!-- <div class="card p-5"> -->
+            <h5 class="card-title">Make a new survey</h5>
+            <form action="/questions" method="get">
+                <div class="my-3">
+                    <label for="title" class="form-label">Survey Name:</label>
+                    <input type="text" class="form-control" id="title" name="title">
+                </div>
+                <div class="mb-3">
+<!--                    <a href ="/questions">-->
+                        <input type = "submit"/>
+<!--                    <button href="/questions" type="submit" class="btn btn-primary">Create Survey</button>-->
+<!--                    </a>-->
+                </div>
+            </form>
+            <!-- </div> -->
+        </div>
+    </div>
+</div>
+
+</div>
+</div>
+<!-- <script type="text/javascript">
+  document.getElementById("yesResult").onclick = function() {
+     document.getElementById("optionResult").style.display = "none";
+     document.getElementById("resultAvailable").style.display = "block";
+
+
+        }
+  document.getElementById("noResult").onclick = function() {
+     document.getElementById("resultAvailable").style.display = "none";
+        }
+</script> -->
+</body>
+</html>
