+++ src/main/resources/templates/monkeyform.html
@@ -109,7 +109,7 @@
         </div>
     </div>
 
-    <p><input onclick="addItem()" type="submit" value="Submit Survey"/></p>
+    <p><input onclick="addItem()" type="submit" value="Submit minisurveymonkey.accessingdatajpa.Survey"/></p>
 </form>
 
 </body>
