+++ src/main/resources/templates/questions.html
@@ -14,7 +14,7 @@
 
     </ol>
 
-    <form action="" method="post" id="form">
+    <form method="post" id="form">
         <div>
             <p>What is your question?</p>
             <input type="text" id="quest"/>
