+++ src/main/resources/templates/starting.html
@@ -8,7 +8,7 @@
 <body>
     <h1>Welcome to Mini-SurveyMonkey!</h1>
 
-    <h2>Want to Create a Survey Today?</h2>
+    <h2>Want to Create a minisurveymonkey.accessingdatajpa.Survey Today?</h2>
     <h3>Click the button below to make your survey.</h3>
 
     <button>
